add SPI动态配置序列化器

+++ common/src/main/java/com/han/rpc/common/service/UserService.java
@@ -18,7 +18,7 @@ public interface UserService {
      * 默认方法-获取数字
      * @return
      */
-    default short getNumber() {
+    default int getNumber() {
         return 1;
     }
 }
